% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOstats.R
\name{go_analysis}
\alias{go_analysis}
\title{Run GOstats functional enrichment analysis for a list of genes.}
\usage{
go_analysis(genes, universe, annotation_package, ontologies = c("BP",
  "CC", "MF"), min_size = 3, mdlinks = TRUE, p_value = 0.01)
}
\arguments{
\item{genes}{The genes of interest; ie the balls drawn from the urn in the
classical hypergeometric analogy.}

\item{universe}{The background set of genes. This corresponds to the bag
of balls in the classical hypergeometric analogy.}

\item{annotation_package}{An organism annotation package used by the GOstats
functions internally to annotate genes with GO terms and to traverse the GO 
hierarchy. eg, for humans, org.Hs.eg.db would be suitable and would require
Entrez IDs for the \code{genes} and \code{universe} arguments.}

\item{ontologies}{The ontologies to test. Options are "BP" 
(biological process), "CC" (cellular component) and "MF" 
(molecular function).}

\item{min_size}{The minimum size of GO terms to include. It is advised to set
this to at least 2; otherwise you will observe quite a few terms which are 
comprised of only a single gene.}

\item{mdlinks}{Should the GO IDs be hyperlinked using Markdown in the output
table?}

\item{p_value}{p-value cutoff. This is important to set a priori, as it 
affects the conditional algorithm.}
}
\description{
Runs the conditional GOstats algorithm for a set of genes in a given 
universe. This function runs the algorithm for multiple ontologies, and returns 
a table of merged results.
}
