% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{rvallejos}
\alias{rvallejos}
\title{rvallejos: The group R package for the Vallejos group.}
\description{
This package contains a lot of TODO!
}
\section{Enrichment analysis}{

GOstats - use this.
}

\section{Utility}{

\link{fix_dates} - fix messy inconsistent dates entered via a free-text box
}

