% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{rvallejos}
\alias{rvallejos}
\alias{rvallejos-package}
\title{rvallejos: The group R package for the Vallejos group.}
\description{
This package contains a lot of TODO!
}
\section{Enrichment analysis}{

GOstats - use this.
}

